Validaciones

+++ CRM-master/C R M/Controllers/PublicidadsController.cs
@@ -11,6 +11,7 @@ using C_R_M.Models;
 
 namespace C_R_M.Controllers
 {
+    [PermisoAttribute]
     public class PublicidadsController : Controller
     {
         private CRMEntities db = new CRMEntities();
@@ -18,18 +19,22 @@ namespace C_R_M.Controllers
         // GET: Publicidads
         public async Task<ActionResult> Index()
         {
-            var publicidad = db.Publicidad.Include(p => p.Empresa1).Include(p => p.EstadodeCuenta).Include(p => p.MedioPublicitario);
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
+            var publicidad = db.Publicidad.Include(p => p.Empresa).Include(p => p.Empresa.EstadodeCuenta).Include(p => p.MedioPublicitario);
             return View(await publicidad.ToListAsync());
         }
 
         // GET: Publicidads/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Publicidad publicidad = await db.Publicidad.FindAsync(id);
+            Publicidad publicidad = await db.Publicidad.Include(p => p.Empresa).Include(p => p.Empresa.EstadodeCuenta).Include(p => p.MedioPublicitario).FirstAsync(p => p.Id_Publicidad == id.Value);
             if (publicidad == null)
             {
                 return HttpNotFound();
@@ -40,10 +45,12 @@ namespace C_R_M.Controllers
         // GET: Publicidads/Create
         public ActionResult Create()
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre");
-            ViewBag.Credito_Disponible = new SelectList(db.EstadodeCuenta, "Id_Estado", "Id_Estado");
             ViewBag.Medio = new SelectList(db.MedioPublicitario, "Id_Medio_Publicitario", "Nombre");
-            return View();
+            Publicidad publicidad = new Publicidad { Fecha_Caducidad = DateTime.Now, Fecha_Inicio = DateTime.Now };
+            return View(publicidad);
         }
 
         // POST: Publicidads/Create
@@ -51,17 +58,33 @@ namespace C_R_M.Controllers
         // más información vea https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include = "Id_Publicidad,Medio,Empresa,Credito_Disponible,Fecha_Inicio,Fecha_Caducidad,Costo")] Publicidad publicidad)
+        public async Task<ActionResult> Create([Bind(Include = "Id_Publicidad,Medio,Id_empresa,Fecha_Inicio,Fecha_Caducidad,Costo")] Publicidad publicidad)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (ModelState.IsValid)
             {
                 db.Publicidad.Add(publicidad);
+                EstadodeCuenta est = null;
+                try
+                {
+                    est = await db.EstadodeCuenta.FirstAsync(es => es.Id_Empresa == publicidad.Id_empresa);
+                }
+                catch 
+                {
+                    est = null;
+                }
+                
+                if (est == null)
+                    db.EstadodeCuenta.Add(new EstadodeCuenta { Id_Empresa = publicidad.Id_empresa, Credito_Disponible = ((decimal)((publicidad.Costo != null) ? (0 - publicidad.Costo.Value) : 0)) });
+                else
+                {
+                    est.Credito_Disponible = Decimal.Round(est.Credito_Disponible - (decimal)(publicidad.Costo.HasValue ? publicidad.Costo.Value : 0), 2);
+                }
                 await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
-
             ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", publicidad.Empresa);
-            ViewBag.Credito_Disponible = new SelectList(db.EstadodeCuenta, "Id_Estado", "Id_Estado", publicidad.Credito_Disponible);
             ViewBag.Medio = new SelectList(db.MedioPublicitario, "Id_Medio_Publicitario", "Nombre", publicidad.Medio);
             return View(publicidad);
         }
@@ -69,6 +92,8 @@ namespace C_R_M.Controllers
         // GET: Publicidads/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -78,8 +103,6 @@ namespace C_R_M.Controllers
             {
                 return HttpNotFound();
             }
-            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", publicidad.Empresa);
-            ViewBag.Credito_Disponible = new SelectList(db.EstadodeCuenta, "Id_Estado", "Id_Estado", publicidad.Credito_Disponible);
             ViewBag.Medio = new SelectList(db.MedioPublicitario, "Id_Medio_Publicitario", "Nombre", publicidad.Medio);
             return View(publicidad);
         }
@@ -89,16 +112,24 @@ namespace C_R_M.Controllers
         // más información vea https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include = "Id_Publicidad,Medio,Empresa,Credito_Disponible,Fecha_Inicio,Fecha_Caducidad,Costo")] Publicidad publicidad)
+        public async Task<ActionResult> Edit([Bind(Include = "Id_Publicidad,Medio,Id_empresa,Fecha_Inicio,Fecha_Caducidad,Costo")] Publicidad publicidad)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (ModelState.IsValid)
             {
-                db.Entry(publicidad).State = EntityState.Modified;
+                Publicidad publicidadtem = await db.Publicidad.FindAsync(publicidad.Id_Publicidad);
+                EstadodeCuenta est = await db.EstadodeCuenta.FirstAsync(es => es.Id_Empresa == publicidad.Id_empresa);
+                if (publicidadtem.Costo != null)
+                    est.Credito_Disponible += (decimal)publicidadtem.Costo;
+                est.Credito_Disponible -= (decimal)((publicidad.Costo != null) ? (publicidad.Costo.Value) : 0);
+                publicidadtem.Costo = publicidad.Costo;
+                publicidadtem.Medio = publicidad.Medio;
+                publicidadtem.Fecha_Caducidad = publicidad.Fecha_Caducidad;
+                publicidadtem.Fecha_Inicio = publicidad.Fecha_Inicio;
                 await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
-            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", publicidad.Empresa);
-            ViewBag.Credito_Disponible = new SelectList(db.EstadodeCuenta, "Id_Estado", "Id_Estado", publicidad.Credito_Disponible);
             ViewBag.Medio = new SelectList(db.MedioPublicitario, "Id_Medio_Publicitario", "Nombre", publicidad.Medio);
             return View(publicidad);
         }
@@ -106,11 +137,13 @@ namespace C_R_M.Controllers
         // GET: Publicidads/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Publicidad publicidad = await db.Publicidad.FindAsync(id);
+            Publicidad publicidad = await db.Publicidad.Include(p => p.Empresa).Include(p => p.Empresa.EstadodeCuenta).Include(p => p.MedioPublicitario).FirstAsync(p => p.Id_Publicidad == id.Value);
             if (publicidad == null)
             {
                 return HttpNotFound();
@@ -123,7 +156,12 @@ namespace C_R_M.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             Publicidad publicidad = await db.Publicidad.FindAsync(id);
+            EstadodeCuenta est = await db.EstadodeCuenta.FirstAsync(es => es.Id_Empresa == publicidad.Id_empresa);
+            if (publicidad.Costo != null)
+                est.Credito_Disponible += (decimal)publicidad.Costo;
             db.Publicidad.Remove(publicidad);
             await db.SaveChangesAsync();
             return RedirectToAction("Index");
