+++ CRM-master/C R M/Controllers/RecordatoriosController.cs
@@ -11,6 +11,7 @@ using C_R_M.Models;
 
 namespace C_R_M.Controllers
 {
+    [PermisoAttribute]
     public class RecordatoriosController : Controller
     {
         private CRMEntities db = new CRMEntities();
@@ -18,18 +19,22 @@ namespace C_R_M.Controllers
         // GET: Recordatorios
         public async Task<ActionResult> Index()
         {
-            var recordatorio = db.Recordatorio.Include(r => r.Empresa1).Include(r => r.Recordar);
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
+            var recordatorio = db.Recordatorio.Include(r => r.Empresa);
             return View(await recordatorio.ToListAsync());
         }
 
         // GET: Recordatorios/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Recordatorio recordatorio = await db.Recordatorio.FindAsync(id);
+            Recordatorio recordatorio = await db.Recordatorio.Include(r => r.Empresa).FirstAsync(r=>r.Id_Recordatorio == id.Value);
             if (recordatorio == null)
             {
                 return HttpNotFound();
@@ -40,9 +45,11 @@ namespace C_R_M.Controllers
         // GET: Recordatorios/Create
         public ActionResult Create()
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre");
-            ViewBag.Id_Recordar = new SelectList(db.Recordar, "Id_Recordar", "Descripción");
-            return View();
+            Recordatorio recordatorio = new Recordatorio { Fecha = DateTime.Now };
+            return View(recordatorio);
         }
 
         // POST: Recordatorios/Create
@@ -50,8 +57,10 @@ namespace C_R_M.Controllers
         // más información vea https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include = "Id_Recordatorio,Tipo,Fecha,Hora,Minutos,Abreviatura,Detalle,Empresa,Id_Recordar")] Recordatorio recordatorio)
+        public async Task<ActionResult> Create([Bind(Include = "Id_Recordatorio,Tipo,Fecha,Hora,Abreviatura,Detalle,Id_empresa,Mensaje")] Recordatorio recordatorio)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (ModelState.IsValid)
             {
                 db.Recordatorio.Add(recordatorio);
@@ -59,14 +68,15 @@ namespace C_R_M.Controllers
                 return RedirectToAction("Index");
             }
 
-            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", recordatorio.Empresa);
-            ViewBag.Id_Recordar = new SelectList(db.Recordar, "Id_Recordar", "Descripción", recordatorio.Id_Recordar);
+            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", recordatorio.Id_empresa);
             return View(recordatorio);
         }
 
         // GET: Recordatorios/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -76,8 +86,7 @@ namespace C_R_M.Controllers
             {
                 return HttpNotFound();
             }
-            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", recordatorio.Empresa);
-            ViewBag.Id_Recordar = new SelectList(db.Recordar, "Id_Recordar", "Descripción", recordatorio.Id_Recordar);
+            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", recordatorio.Id_empresa);
             return View(recordatorio);
         }
 
@@ -86,22 +95,25 @@ namespace C_R_M.Controllers
         // más información vea https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include = "Id_Recordatorio,Tipo,Fecha,Hora,Minutos,Abreviatura,Detalle,Empresa,Id_Recordar")] Recordatorio recordatorio)
+        public async Task<ActionResult> Edit([Bind(Include = "Id_Recordatorio,Tipo,Fecha,Hora,Abreviatura,Detalle,Id_empresa,Mensaje")] Recordatorio recordatorio)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (ModelState.IsValid)
             {
                 db.Entry(recordatorio).State = EntityState.Modified;
                 await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
-            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", recordatorio.Empresa);
-            ViewBag.Id_Recordar = new SelectList(db.Recordar, "Id_Recordar", "Descripción", recordatorio.Id_Recordar);
+            ViewBag.Empresa = new SelectList(db.Empresa, "Id_Empresa", "Nombre", recordatorio.Id_empresa);
             return View(recordatorio);
         }
 
         // GET: Recordatorios/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -119,10 +131,27 @@ namespace C_R_M.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
             Recordatorio recordatorio = await db.Recordatorio.FindAsync(id);
             db.Recordatorio.Remove(recordatorio);
             await db.SaveChangesAsync();
             return RedirectToAction("Index");
+        }
+        [AllowAnonymous]
+        public async Task<ActionResult> RecordatoriosDia()
+        {
+            if (AccountController.Account.GetUser == null)
+                return RedirectPermanent("Login/Index");
+            var recordatorio = await db.Recordatorio.Include(r => r.Empresa).ToListAsync();
+            return View(recordatorio.FindAll(r => r.Fecha.Date == DateTime.Now.Date));
+        }
+        [AllowAnonymous]
+        public static int RecordatoriosCount()
+        {
+            CRMEntities db = new CRMEntities();
+            var recordatorio = db.Recordatorio.ToList();
+            return recordatorio.FindAll(r => r.Fecha.Date == DateTime.Now.Date).Count;
         }
 
         protected override void Dispose(bool disposing)
